refactor: reorganize Modal and related components
